# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentX Marketplace API",
    "version": "0.2.0",
    "description": "AI agent commerce layer — discover, compare, and transact with verified service providers. Supports x402 payments on Solana, Base, Ethereum, and Bitcoin.",
    "contact": {
      "email": "info@agentx.sale",
      "url": "https://agentx.sale"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://agentx.sale/api",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/challenge": {
      "get": {
        "summary": "Get PoW challenge",
        "description": "Returns a proof-of-work challenge for agent authentication.",
        "operationId": "getChallenge",
        "responses": {
          "200": {
            "description": "Challenge details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Challenge"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Verify challenge solution",
        "description": "Submit a solved PoW challenge. Returns session token on success.",
        "operationId": "verifyChallenge",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challenge_id": { "type": "string" },
                  "data": { "type": "string" },
                  "nonce": { "type": "integer" },
                  "prefix": { "type": "string" }
                },
                "required": ["challenge_id", "data", "nonce", "prefix"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeVerify"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Challenge": {
        "type": "object",
        "properties": {
          "challenge_id": { "type": "string" },
          "data": { "type": "string" },
          "prefix": { "type": "string" },
          "difficulty": { "type": "integer" },
          "algorithm": { "type": "string", "example": "SHA256" },
          "expires_at": { "type": "string", "format": "date-time" },
          "instructions": { "type": "string" }
        }
      },
      "ChallengeVerify": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "verified" },
          "message": { "type": "string" },
          "session_token": { "type": "string" },
          "valid_until": { "type": "string", "format": "date-time" },
          "challenge_id": { "type": "string" },
          "nonce_found": { "type": "integer" },
          "hash": { "type": "string" }
        }
      },
...
    }
  }
}